refactoring and seperating numbers, duration and dattime extensions
