+++ example/duration_example.dart
@@ -1,0 +1,7 @@
+import 'package:time/time.dart';
+
+void main() {
+  print(7.days.inWeeks);
+  print(7.days.fromNow);
+  print(7.days.ago);
+}
