+++ lib/src/duration_extensions.dart
@@ -1,0 +1,18 @@
+part of time;
+
+extension DurationTimeExtension on Duration {
+  static const int daysPerWeek = 7;
+  static const int nanosecondsPerMicrosecond = 1000;
+
+  /// Returns the representation in weeks
+  int get inWeeks => (inDays / daysPerWeek).ceil();
+
+  /// Adds the Duration to the current DateTime and returns a DateTime in the future
+  DateTime get fromNow => DateTime.now() + this;
+
+  @Deprecated('Use fromNow instead. Will be removed in 2.0.0')
+  DateTime get later => fromNow;
+
+  /// Subtracts the Duration from the current DateTime and returns a DateTime in the past
+  DateTime get ago => DateTime.now() - this;
+}
