+++ lib/time.dart
@@ -1,3 +1,5 @@
 library time;
 
-export './src/extensions.dart';
+part 'src/datetime_extensions.dart';
+part 'src/num_extensions.dart';
+part 'src/duration_extensions.dart';
